Finetuning/MiniCPM and InternVL/requirements.txt
# PyTorch ecosystem with CUDA 11.8 support
--index-url https://download.pytorch.org/whl/cu118
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0

# Core dependencies
packaging==23.2
numpy==1.24.4
Pillow==10.1.0
matplotlib==3.7.4
opencv-python-headless==4.5.5.64

# Main requirements
addict==2.4.0
editdistance==0.6.2
einops==0.8.0
fairscale==0.4.0
jsonlines==4.0.0
markdown2==2.4.10
more_itertools==10.1.0
nltk==3.8.1
openpyxl==3.1.2
sacrebleu==2.3.2
seaborn==0.13.0
shortuuid==1.0.11
spacy==3.7.2
timm==0.9.10
tqdm==4.66.1
protobuf==4.25.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
sentencepiece==0.1.99
accelerate==0.30.1
socksio==1.0.0
gradio
gradio_client
decord
aiosignal
tensorboard
deepspeed==0.12.3
transformers==4.44.2
librosa==0.9.0
soundfile==0.12.1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
peft==0.14.0
ninja==1.11.1.1
flash-attn
moviepy
http://thunlp.oss-cn-qingdao.aliyuncs.com/multi_modal/never_delete/modelscope_studio-0.4.0.9-py3-none-any.whl
https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.8/flash_attn-2.5.8+cu118torch2.2cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
datasets
